Fix URL for a dependency (#5)
The URL for angular-legacy-sortablejs-maintained is incorrect. According
to the manual for package.json[^1], the URL would have to start with
`git+https`. Otherwise, it could be interpreted as a URL that points to
a tarball. Yarn and NPM do not have any issue with this but yarn2nix[^2]
does fail to build a package because of this. And this might affect
other tools as well.

Since the URL in question points to GitHub, this commit fixes the
problem by using the abbreviated form for GitHub URLs.

[^1]: <https://docs.npmjs.com/cli/v8/configuring-npm/package-json#git-urls-as-dependencies>
[^2]: <https://nixos.org/manual/nixpkgs/stable/#javascript-tool-specific>